refactor: move `umount` to the common executable
