--- umount/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "umount"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-libc = "*"
